Frontend- sidebar component
