+++ itms.client/itms.client.esproj
@@ -10,4 +10,22 @@
   <ItemGroup>
     <Folder Include="src\app\shared\components\NewFolder\" />
   </ItemGroup>
+  <ItemGroup>
+    <TypeScriptConfiguration Remove="src\app\shared\components\sidebar-new\**" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="src\app\shared\components\sidebar-new\**" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="src\app\shared\components\sidebar-new\sidebar-new.component.ts" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="src\app\shared\components\sidebar-new\sidebar-new.component.spec.ts" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="src\app\shared\components\sidebar-new\sidebar-new.component.html" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="src\app\shared\components\sidebar-new\sidebar-new.component.css" />
+  </ItemGroup>
 </Project>
