+++ itms.client/src/app/app.component.html
@@ -1,11 +1,13 @@
 <app-navbar></app-navbar>
-<app-new-sidebar></app-new-sidebar>
-<main>
-  <div class="container-fluid">
-    <div class="row">
-      <div class="col-md-12">
-        <router-outlet></router-outlet>
+<div style="display: flex;">
+  <app-new-sidebar></app-new-sidebar>
+ </div>
+  <main>
+    <div class="container-fluid">
+      <div class="row">
+        <div class="col-md-12">
+          <router-outlet></router-outlet>
+        </div>
       </div>
     </div>
-  </div>
-</main>
+  </main>
