+++ itms.client/src/app/app.component.scss
@@ -1,13 +1,12 @@
-app-sidebar {
+app-new-sidebar {
   position: fixed;
   left: 0rem;
-  height: 776px;
   z-index: 4;
   box-shadow: 5px 0px 5px -5px rgba(0, 0, 0, 0.2);
   margin-right: 2%;
   background-color: white;
   width: 60px;
-  top:56px;
+  
 }
 
 main {
