+++ itms.client/src/app/app.module.ts
@@ -3,7 +3,6 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { SidebarComponent } from './shared/components/sidebar/sidebar.component';
 import { NavbarComponent } from './shared/components/navbar/navbar.component';
 import { AvatarModule } from 'ngx-avatars';
 import { HttpClientModule } from '@angular/common/http';
@@ -14,13 +13,14 @@ import { EmployeeComponent } from './employee/employee.component';
 import { LicenseComponent } from './license/license.component';
 import { HistoryComponent } from './history/history.component';
 import { AddAssetComponent } from './add-asset/add-asset.component';
-import { NewSidebarComponent } from './new-sidebar/new-sidebar.component';
+import { ProfileIconComponent } from './shared/components/profile-icon/profile-icon.component';
+import { NewSidebarComponent } from './shared/components/new-sidebar/new-sidebar.component';
+// Use SidebarNewComponent in your code
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    SidebarComponent,
     NavbarComponent,
     DashboardComponent,
     CategoryComponent,
@@ -30,6 +30,7 @@ import { NewSidebarComponent } from './new-sidebar/new-sidebar.component';
     HistoryComponent,
     AddAssetComponent,
     NewSidebarComponent,
+    ProfileIconComponent,
   ],
   imports: [
     BrowserModule,
