--- itms.client/src/app/new-sidebar/new-sidebar.component.html
@@ -1,14 +1,0 @@
-<!-- sidebar.component.html -->
-<div class="side-nav-content" [ngClass]="{'collapsed': isSidebarCollapsed}">
-  <ul class="nav-list">
-    <li class="nav-list-item" *ngFor="let item of list">
-      <a class="hover-container side-nav-icons" [routerLink]="['/', item.name.toLowerCase()]">
-        <img src="{{ item.icon }}" />
-        <span class="name" [class.visible]="!isSidebarCollapsed">{{ item.name }}</span>
-      </a>
-    </li>
-    <button class="collapse-icon" (click)="toggleSidebar()">
-      <img src="../../../../assets/icons/Group.svg" />
-    </button>
-  </ul>
-</div>
