+++ itms.client/src/app/shared/components/navbar/navbar.component.scss
@@ -17,23 +17,21 @@
 }
 
 .my-navbar {
-  box-shadow: 0px 1px 5px 0px rgba(0, 0, 0, 0.2);
   position: fixed;
   top: 0;
   left: 0;
-  width: 100%;
   z-index: 1000;
-  height: 56px !important;
+  display: flex;
+  width: 80rem;
+  height: 3.5rem !important;
+  padding: 0rem 1.25rem 0rem 1rem;
+  justify-content: space-between;
+  align-items: center;
+  flex-shrink: 0;
+  background: var(--White, #FFF);
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.15);
+  margin-bottom: 3.5rem;
 }
-
-/*.toggle_box {
-  display: block;
-  background-color: #e9eceb;
-  border-radius: 1000px;
-  cursor: pointer;
-  position: relative;
-  padding: 2%;
-}*/
 
 .btn-check:checked + .btn {
   background-color: #28519e;
@@ -47,48 +45,53 @@
 }
 .dropdown-both:hover {
   background-color: lightgrey;
-  /* Remove default margin */
   display: flex;
   gap: 8px;
   text-align: center;
+  width:100%;
   align-items: center;
+  margin-left:0% !important;
 }
 
 .dropdown-btn {
   display: flex;
-  margin-left:10px;
+  margin-left: 10px;
   align-items: center;
   gap: 8px;
-  text-align:center;
+  text-align: center;
+  
 }
+/*.dropdown-icn{
+    padding:0.38rem 0.88rem 0.38rem 0.75rem;
+}*/
 .additional-options {
-  top: 45px;
-  width: 161px;
-  border-radius: 4px;
+  border-radius: 0.25rem;
+  top: 3rem;
   background: #FFF;
   box-shadow: 0px 4px 20px 0px rgba(40, 52, 73, 0.15);
- padding:10px 0px 10px 0px;
+  display: flex;
+  width: 10.0625rem;
+  padding: 0.75rem 0rem;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.75rem;
 }
 .dropdown-both {
   display: flex;
-  gap: 8px; /*
-  text-align:center;
-      */
-  align-items:center;
-  padding: 0px 5px 0px 5px;
+  align-items: center;
+  gap: 0.5rem;
+  align-self: stretch;
+  background: #FFF;
+  color: var(--Primary-Gray, #343435);
+  padding: 0.38rem 0.88rem 0.38rem 0.75rem;
+  font-family: Roboto;
+  font-size: 0.9375rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 1.5rem; /* 160% */
+  width: 100%;
+  margin-left: 0% !important;
 }
-
-/*.selected {
-  position: absolute;
-  top: 1px;
-  left: 1px;
-  width: calc(50% - 1px);
-  height: calc(100% - 2px);
-  border-radius: 20px;
-  transition: transform 0.2s;
-  background-color: var(--blue);
-}
-*/
 
 input[type=radio]:checked + .off-label + .selected {
   transform: translateX(100%);
@@ -140,19 +143,19 @@ input[type=radio]:checked + .off-label + .selected {
   /* Toggle Shadow */
   box-shadow: 0px 1px 4px 0px rgba(40, 81, 158, 0.30);
   box-shadow: rgba(0, 0, 0, 0.16) 0px 1px 4px;
-  transition: 0.25s ease-in;
+  transition: 0.25s ease-in-out;
 }
 
 .inputf input:nth-child(1):checked ~ .slider {
-  left: 2px;
+  left: 3px;
 }
 
 .inputf input:nth-child(3):checked ~ .slider {
-  left: 67px;
+  left: 68.5px;
 }
 
 .inputf input:nth-child(5):checked ~ .slider {
-  left: 132px;
+  left: 135px;
 }
 
 .inputf input:nth-child(1):checked + label {
@@ -178,15 +181,16 @@ input[type=radio]:checked + .off-label + .selected {
 }
 .dropdown-main {
   display: flex;
-  padding: 8px 8px 8px 12px;
   justify-content: center;
   align-items: center;
-  gap: 8px;
-  border-radius: 4px;
+  border-radius: 0.25rem;
   background: var(--Primary-Blue, #28519E);
   color: var(--Primary-White, #FFF);
-  font-size: 16px;
+  font-family: Roboto;
+  font-size: 1rem;
   font-style: normal;
-  font-weight: 400;
+  font-weight: 500;
   line-height: normal;
+  padding: 0.5rem 0.5rem 0.5rem 0.75rem;
+  gap: 0.5rem;
 }
