+++ itms.client/src/app/shared/components/profile-icon/profile-icon.component.css
@@ -1,0 +1,6 @@
+.avatar-dropdown {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  cursor: pointer;
+}
