+++ itms.client/src/app/shared/components/profile-icon/profile-icon.component.html
@@ -1,0 +1,4 @@
+<div class="avatar-dropdown">
+  <ngx-avatars name="John Doe" [textSizeRatio]=2.7 bgColor="#b8dde1" fgColor="#264f9d" [round]="true"></ngx-avatars>
+  <img src="../../../../assets/icons/arrow.svg">
+</div>
