+++ itms.client/src/app/shared/components/profile-icon/profile-icon.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ProfileIconComponent } from './profile-icon.component';
+
+describe('ProfileIconComponent', () => {
+  let component: ProfileIconComponent;
+  let fixture: ComponentFixture<ProfileIconComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [ProfileIconComponent]
+    });
+    fixture = TestBed.createComponent(ProfileIconComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
