+++ itms.client/src/app/shared/components/profile-icon/profile-icon.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-profile-icon',
+  templateUrl: './profile-icon.component.html',
+  styleUrls: ['./profile-icon.component.css']
+})
+export class ProfileIconComponent {
+
+}
